testtools/tests/test_deferredruntest.py
# Copyright (c) 2010 Jonathan M. Lange. See LICENSE for details.

"""Tests for the DeferredRunTest single test execution logic."""

from testtools import (
    TestCase,
    )
from testtools.deferredruntest import SynchronousDeferredRunTest
from testtools.tests.helpers import ExtendedTestResult
from testtools.matchers import Equals

from twisted.internet import defer


class TestSynchronousDeferredRunTest(TestCase):

    def test_success(self):
        class SomeCase(TestCase):
            def test_success(self):
                return defer.succeed(None)
        test = SomeCase('test_success')
        runner = SynchronousDeferredRunTest(test)
        result = ExtendedTestResult()
        runner.run(result)
        self.assertThat(
            result._events, Equals([
                ('startTest', test),
                ('addSuccess', test),
                ('stopTest', test)]))

    def test_failure(self):
        class SomeCase(TestCase):
            def test_failure(self):
                return defer.maybeDeferred(self.fail, "Egads!")
        test = SomeCase('test_failure')
        runner = SynchronousDeferredRunTest(test, test.exception_handlers)
        result = ExtendedTestResult()
        runner.run(result)
        self.assertThat(
            [event[:2] for event in result._events], Equals([
                ('startTest', test),
                ('addFailure', test),
                ('stopTest', test)]))

    def test_setUp_followed_by_test(self):
        class SomeCase(TestCase):
            def setUp(self):
                super(SomeCase, self).setUp()
                return defer.succeed(None)
            def test_failure(self):
                return defer.maybeDeferred(self.fail, "Egads!")
        test = SomeCase('test_failure')
        runner = SynchronousDeferredRunTest(test, test.exception_handlers)
        result = ExtendedTestResult()
        runner.run(result)
        self.assertThat(
            [event[:2] for event in result._events], Equals([
                ('startTest', test),
                ('addFailure', test),
                ('stopTest', test)]))


def test_suite():
    from unittest import TestLoader
    return TestLoader().loadTestsFromName(__name__)
